requirements.txt
aiohttp
pydantic
aiocsv
aiofiles
python-dotenv
gspread
google-auth
fastapi
uvicorn
python-multipart
pyinstaller
